chore(reannounce): move CanMatchTorrents test beside the model and document the Monitor All default
